+++ crates/swc_css_ast/src/style_rule.rs
@@ -1,5 +1,32 @@
 use crate::{AtRule, DashedIdent, Ident, SelectorList, Tokens, Value};
+use is_macro::Is;
 use swc_common::{ast_node, Span};
+
+#[ast_node("Stylesheet")]
+pub struct Stylesheet {
+    pub span: Span,
+    pub rules: Vec<Rule>,
+}
+
+#[ast_node]
+#[derive(Is)]
+pub enum Rule {
+    #[tag("QualifiedRule")]
+    QualifiedRule(QualifiedRule),
+
+    #[tag("Tokens")]
+    Invalid(Tokens),
+
+    #[tag("*")]
+    AtRule(AtRule),
+}
+
+#[ast_node("QualifiedRule")]
+pub struct QualifiedRule {
+    pub span: Span,
+    pub prelude: SelectorList,
+    pub block: Block,
+}
 
 #[ast_node("SimpleBlock")]
 pub struct SimpleBlock {
@@ -8,13 +35,6 @@ pub struct SimpleBlock {
     // its value initially set to an empty list.
     pub name: char,
     pub value: Vec<Value>,
-}
-
-#[ast_node("QualifiedRule")]
-pub struct QualifiedRule {
-    pub span: Span,
-    pub prelude: SelectorList,
-    pub block: Block,
 }
 
 #[ast_node("Block")]
@@ -33,14 +53,6 @@ pub enum DeclarationBlockItem {
     AtRule(AtRule),
 }
 
-#[ast_node]
-pub enum DeclarationProperty {
-    #[tag("Ident")]
-    Ident(Ident),
-    #[tag("DashedIdent")]
-    DashedIdent(DashedIdent),
-}
-
 #[ast_node("Declaration")]
 pub struct Declaration {
     pub span: Span,
@@ -48,4 +60,12 @@ pub struct Declaration {
     pub value: Vec<Value>,
     /// The span includes `!`
     pub important: Option<Span>,
+}
+
+#[ast_node]
+pub enum DeclarationProperty {
+    #[tag("Ident")]
+    Ident(Ident),
+    #[tag("DashedIdent")]
+    DashedIdent(DashedIdent),
 }
