+++ crates/swc_css_parser/tests/fixture.rs
@@ -281,7 +281,6 @@ macro_rules! mtd {
 impl Visit for SpanVisualizer<'_> {
     mtd!(AtRule, visit_at_rule);
     mtd!(AtSelector, visit_at_selector);
-    mtd!(AtTextValue, visit_at_text_value);
     mtd!(AttrSelector, visit_attr_selector);
     mtd!(BinValue, visit_bin_value);
     mtd!(ClassSelector, visit_class_selector);
