+++ crates/swc_css_parser/tests/recovery/value/at-keyword/1/output.json
@@ -71,44 +71,60 @@
         },
         "value": [
           {
-            "type": "Declaration",
+            "type": "Tokens",
             "span": {
               "start": 4,
-              "end": 16,
+              "end": 17,
               "ctxt": 0
             },
-            "property": {
-              "type": "Identifier",
-              "span": {
-                "start": 4,
-                "end": 9,
-                "ctxt": 0
-              },
-              "value": "value",
-              "raw": "value"
-            },
-            "value": [
+            "tokens": [
               {
-                "type": "AtTextValue",
                 "span": {
-                  "start": 11,
-                  "end": 16,
+                  "start": 4,
+                  "end": 9,
                   "ctxt": 0
                 },
-                "name": {
-                  "type": "Identifier",
-                  "span": {
-                    "start": 11,
-                    "end": 16,
-                    "ctxt": 0
-                  },
-                  "value": ",x",
-                  "raw": "\\2cx"
+                "token": {
+                  "Ident": {
+                    "value": "value",
+                    "raw": "value"
+                  }
+                }
+              },
+              {
+                "span": {
+                  "start": 9,
+                  "end": 10,
+                  "ctxt": 0
                 },
-                "block": null
+                "token": "Colon"
+              },
+              {
+                "span": {
+                  "start": 10,
+                  "end": 11,
+                  "ctxt": 0
+                },
+                "token": {
+                  "WhiteSpace": {
+                    "value": " "
+                  }
+                }
+              },
+              {
+                "span": {
+                  "start": 11,
+                  "end": 17,
+                  "ctxt": 0
+                },
+                "token": {
+                  "AtKeyword": {
+                    "value": "x,",
+                    "raw": "x\\2c "
+                  }
+                }
               }
-            ],
-            "important": null
+            ]
           }
         ]
       }
