+++ crates/swc_css_parser/tests/recovery/value/at-keyword/1/output.swc-stderr
@@ -1,0 +1,6 @@
+error: Expected Declaration value
+ --> $DIR/tests/recovery/value/at-keyword/1/input.css:1:12
+  |
+1 | a { value: @x\2c }
+  |            ^^^^^^
+
