+++ crates/swc_css_parser/tests/recovery/value/at-keyword/3/output.json
@@ -2,7 +2,7 @@
   "type": "Stylesheet",
   "span": {
     "start": 0,
-    "end": 17,
+    "end": 23,
     "ctxt": 0
   },
   "rules": [
@@ -10,7 +10,7 @@
       "type": "QualifiedRule",
       "span": {
         "start": 0,
-        "end": 17,
+        "end": 23,
         "ctxt": 0
       },
       "prelude": {
@@ -66,49 +66,77 @@
         "type": "Block",
         "span": {
           "start": 2,
-          "end": 17,
+          "end": 23,
           "ctxt": 0
         },
         "value": [
           {
-            "type": "Declaration",
+            "type": "Tokens",
             "span": {
               "start": 4,
-              "end": 15,
+              "end": 22,
               "ctxt": 0
             },
-            "property": {
-              "type": "Identifier",
-              "span": {
-                "start": 4,
-                "end": 9,
-                "ctxt": 0
-              },
-              "value": "value",
-              "raw": "value"
-            },
-            "value": [
+            "tokens": [
               {
-                "type": "AtTextValue",
                 "span": {
-                  "start": 11,
-                  "end": 15,
+                  "start": 4,
+                  "end": 9,
                   "ctxt": 0
                 },
-                "name": {
-                  "type": "Identifier",
-                  "span": {
-                    "start": 11,
-                    "end": 15,
-                    "ctxt": 0
-                  },
-                  "value": "x,",
-                  "raw": "x\\,"
+                "token": {
+                  "Ident": {
+                    "value": "value",
+                    "raw": "value"
+                  }
+                }
+              },
+              {
+                "span": {
+                  "start": 9,
+                  "end": 10,
+                  "ctxt": 0
                 },
-                "block": null
+                "token": "Colon"
+              },
+              {
+                "span": {
+                  "start": 10,
+                  "end": 11,
+                  "ctxt": 0
+                },
+                "token": {
+                  "WhiteSpace": {
+                    "value": " "
+                  }
+                }
+              },
+              {
+                "span": {
+                  "start": 11,
+                  "end": 21,
+                  "ctxt": 0
+                },
+                "token": {
+                  "AtKeyword": {
+                    "value": "keyword",
+                    "raw": "k\\65yword"
+                  }
+                }
+              },
+              {
+                "span": {
+                  "start": 21,
+                  "end": 22,
+                  "ctxt": 0
+                },
+                "token": {
+                  "WhiteSpace": {
+                    "value": " "
+                  }
+                }
               }
-            ],
-            "important": null
+            ]
           }
         ]
       }
